Update getSettingsHtml function to be compatible with Craft 3.5
Craft 3.5 updated the `$view->namespaceInputId` to return an empty string if an empty string was passed to it. This resulted in an Javascript uncaught error on the field settings page as the element with the namespaced id could not be found. Instead, we replace the `namespaceInputId` function with the `Html::namespaceId` helper function to return the value expected by the SEO plugin.